UserDAOFactory эксепшны + ApplicationConfig

+++ src/main/java/Servlets/DeleteServlet.java
@@ -10,15 +10,21 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/delete")
+@WebServlet("/admin/delete")
 public class DeleteServlet extends HttpServlet {
 
     UserService userService = UserService.getInstance();
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        userService.deleteUser(req.getParameter("name"));
-        req.setAttribute("isVisible", true);
-        getServletContext().getRequestDispatcher("/ShowUsersServlet").forward(req, resp);
+        if (req.getAttribute("access") == "false") {
+            req.setAttribute("isVisible", true);
+            getServletContext().getRequestDispatcher("/index.jsp").forward(req, resp);
+        }
+        else {
+            userService.deleteUser(req.getParameter("name"));
+            req.setAttribute("isVisible", true);
+            getServletContext().getRequestDispatcher("/admin").forward(req, resp);
+        }
     }
 }
